refactor: move cli to directory
