+++ package.json
@@ -49,7 +49,7 @@
     }
   },
   "bin": {
-    "libnest": "./dist/cli.js"
+    "libnest": "./dist/cli/index.js"
   },
   "files": [
     "dist"
