+++ src/index.ts
@@ -1,4 +1,4 @@
-import type { ConfigOptions } from './cli.js';
+import type { ConfigOptions } from './cli/index.js';
 
 export function defineConfig(config: ConfigOptions) {
   return config;
